+++ package.json
@@ -1,0 +1,25 @@
+{
+  "name": "steamwebtools",
+  "main": "Gruntfile.js",
+  "dependencies": {
+    "grunt": "^0.4.5"
+  },
+  "devDependencies": {
+    "grunt-contrib-clean": "^1.0.0",
+    "grunt-contrib-uglify": "^0.11.1",
+    "grunt-includes": "^0.5.3"
+  },
+  "scripts": {
+    "test": "echo \"Error: no test specified\" && exit 1"
+  },
+  "repository": {
+    "type": "git",
+    "url": "git+https://github.com/Mr-VIT/SteamWebTools.git"
+  },
+  "author": "Mr-VIT",
+  "license": "GPLv2",
+  "bugs": {
+    "url": "https://github.com/Mr-VIT/SteamWebTools/issues"
+  },
+  "homepage": "http://v1t.su/projects/steam/webtools/"
+}
