+++ src/_version.js
@@ -1,0 +1,2 @@
+// @version		0.5.6
+// @date		2016-02-24
