+++ src/version.js
@@ -1,7 +1,6 @@
 // ==UserScript==
 // @name 	Steam Web Tools
 // @namespace	http://v1t.su/projects/steam/webtools/
-// @version		0.5.6
-// @date		2016-02-24
+//!include _version.js
 // @downloadURL	https://raw.githubusercontent.com/Mr-VIT/SteamWebTools/master/release/steamwebtools.user.js
 // ==/UserScript==
